fixes broken imports caused by moving modules
