+++ client/app/components/_joinGroup/joinGroup.js
@@ -1,7 +1,7 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
 import ngMaterial from "angular-material";
-import GroupService from "../../common/group/group";
+import GroupService from "../../services/group/group";
 import joinGroupComponent from "./joinGroup.component";
 
 let joinGroupModule = angular.module("joinGroup", [
