+++ client/app/components/group/_pickupList/_createPickup/createPickup.js
@@ -1,6 +1,6 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
-import PickupDate from "../../common/pickupDate/pickupDate";
+import PickupDate from "../../../../services/pickupDate/pickupDate";
 import createPickupComponent from "./createPickup.component";
 
 let createPickupModule = angular.module("createPickup", [
