+++ client/app/components/group/group.js
@@ -2,8 +2,8 @@ import angular from "angular";
 import uiRouter from "angular-ui-router";
 import ngMaterial from "angular-material";
 import groupComponent from "./group.component";
-import AuthenticationModule from "../../common/authentication/authentication";
-import groupModule from "../../common/group/group";
+import AuthenticationModule from "../../services/authentication/authentication";
+import groupModule from "../../services/group/group";
 import store from "./store/storeDetail";
 import groupDetail from "./groupDetail/groupDetail";
 import groupEdit from "./groupEdit/groupEdit";
