+++ client/app/components/group/groupDetail/groupDetail.js
@@ -7,7 +7,7 @@ import members from "./members/members";
 import pickups from "./pickups/pickups";
 import stores from "./stores/stores";
 import expandablePanel from "../../_expandablePanel/expandablePanel";
-import GroupService from "../../../common/group/group";
+import GroupService from "../../../services/group/group";
 
 let groupDetailModule = angular.module("groupDetail", [
   uiRouter,
