+++ client/app/components/group/groupDetail/members/_userList/userList.js
@@ -1,7 +1,7 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
 import userListComponent from "./userList.component";
-import User from "../../common/user/user";
+import User from "../../../../../services/user/user.js";
 
 let userListModule = angular.module("userList", [
   uiRouter,
