+++ client/app/components/group/groupDetail/members/members.js
@@ -1,6 +1,6 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
-import userList from "../../../_userList/userList";
+import userList from "./_userList/userList";
 import membersComponent from "./members.component";
 
 let MembersModule = angular.module("members", [
