+++ client/app/components/group/groupDetail/pickups/pickups.js
@@ -1,6 +1,6 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
-import pickupList from "../../../_pickupList/pickupList";
+import pickupList from "../../_pickupList/pickupList";
 import pickupsComponent from "./pickups.component";
 
 let pickupsModule = angular.module("pickups", [
