+++ client/app/components/group/groupEdit/groupEdit.js
@@ -1,8 +1,8 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
-import Authentication from "../../../common/authentication/authentication";
+import Authentication from "../../../services/authentication/authentication";
 import groupEditComponent from "./groupEdit.component";
-import GroupService from "../../../common/group/group";
+import GroupService from "../../../services/group/group";
 import groupEditCreateForm from "../_groupEditCreateForm/groupEditCreateForm";
 
 let groupEditModule = angular.module("groupEdit", [
