+++ client/app/components/home/home.js
@@ -2,7 +2,7 @@ import angular from "angular";
 import uiRouter from "angular-ui-router";
 import homeComponent from "./home.component";
 import logout from "./home.logout";
-import GroupService from "../../common/group/group";
+import GroupService from "../../services/group/group";
 import joinGroup from "../_joinGroup/joinGroup";
 
 let homeModule = angular.module("home", [
