+++ client/app/components/login/login.js
@@ -2,7 +2,7 @@ import angular from "angular";
 import uiRouter from "angular-ui-router";
 import loginComponent from "./login.component";
 import ngMessages from "angular-messages";
-import Authentication from "../../common/authentication/authentication";
+import Authentication from "../../services/authentication/authentication";
 
 let loginModule = angular.module("login", [
   uiRouter,
